add all missing functionality and fixes

+++ server.py
@@ -1,7 +1,6 @@
 import tornado.ioloop
 import tornado.web
 import sqlite3
-import cgitb
 import json
 
 
@@ -45,10 +44,13 @@ class ZespolHandler(tornado.web.RequestHandler):
         c.close()
         
     def delete(self, id):
-        cgitb.enable()
         conn = sqlite3.connect("data/sqlite.db")
         c = conn.cursor()
         try:
+            c.execute("SELECT * FROM zespoly where id = ?",(id,))
+            item = c.fetchone()
+            if item is None:
+                raise Exception()
             c.execute("DELETE FROM zespoly WHERE id = ?", (id,))
             conn.commit()
             self.set_header('Content-Type', 'application/json')
@@ -59,7 +61,6 @@ class ZespolHandler(tornado.web.RequestHandler):
         c.close()
         
     def patch(self,id):
-        cgitb.enable()
         conn = sqlite3.connect("data/sqlite.db")
         c = conn.cursor()
         c.execute("SELECT * FROM zespoly WHERE id = ?", (id,))
@@ -78,11 +79,11 @@ class ZespolHandler(tornado.web.RequestHandler):
             try:
             	nazwa = str(json_data["nazwa"])
             except:
-            	marka = row[1]
+            	nazwa = row[1]
             try:
             	miasto = str(json_data["miasto"])
             except:
-            	model = row[2]
+            	miasto = row[2]
             try:
             	punkty = int(json_data["punkty"])
             except:
@@ -114,13 +115,26 @@ class ZespolyHandler(tornado.web.RequestHandler):
         self.write(output)
         
     def get(self):
-        cgitb.enable()
+        try:
+			page = int(self.get_argument("page"))
+        except:
+			page = 1
+        first = (page - 1)  * 10
+        last = page * 10
+        respone = []
+        
         conn = sqlite3.connect("data/sqlite.db")
         c = conn.cursor()
-        c.execute("select * from zespoly")
+        c.execute("SELECT * FROM zespoly")
+        i = 0
+        j=0
+        for row in c:
+        	if ((j >= first) and (j < last)) and (i<10):
+        		respone.append({"id": row[0], "nazwa": row[1], "miasto": row[2], "punkty": row[3]})
+        		i+=1
+        	j+=1
         self.set_header('Content-Type', 'application/json')
-        for zespol in c:
-            self.write(json.dumps({"id": zespol[0], "nazwa": zespol[1], "miasto": zespol[2], "punkty": zespol[3]},sort_keys=True, indent=4)+"\n")
+        self.write(json.dumps(respone, ensure_ascii=False,sort_keys=True, indent=4))
         conn.commit()
         c.close()
 
@@ -147,7 +161,6 @@ class ZespolyHandler(tornado.web.RequestHandler):
             punkty = str(json_data["punkty"])
         except:
             punkty = 0
-        cgitb.enable()
         conn = sqlite3.connect("data/sqlite.db")
         c = conn.cursor()
         try:
@@ -160,10 +173,7 @@ class ZespolyHandler(tornado.web.RequestHandler):
             self.send_error(500, message=message)
         c.close()
                    
-                   
-                   
     def delete(self):
-        cgitb.enable()
         conn = sqlite3.connect("data/sqlite.db")
         c = conn.cursor()
         try:
@@ -173,6 +183,41 @@ class ZespolyHandler(tornado.web.RequestHandler):
         except:
             message = 'Cannot delete'
             self.send_error(400, message=message)
+        conn.commit()
+        c.close()
+
+    def patch(self):
+        conn = sqlite3.connect("data/sqlite.db")
+        c = conn.cursor()
+        try:
+        	json_data = json.loads(self.request.body)
+        except:
+            message = 'Unable to parse JSON.'
+            self.send_error(400, message=message)
+            return
+        for row in json_data:
+            try:
+            	nazwa = str(row["nazwa"])
+            except:
+                message="Unable to patch object \n"
+                self.send_error(500, message=message)
+            	continue
+            try:
+            	miasto = str(row["miasto"])
+            except:
+            	message="Unable to patch object \n"
+            	self.send_error(500, message=message)
+            	continue
+            try:
+            	punkty = int(row["punkty"])
+            except:
+            	punkty = 0
+            try:
+            	c.execute("UPDATE zespoly SET nazwa=?, miasto=?, punkty=? WHERE id = ?", (nazwa,miasto,punkty,row["id"]))
+            	self.write("OK");
+            except:
+            	message = "Couldn't patch record"
+                self.send_error(500, message=message)
         conn.commit()
         c.close()
